Relocate all simulation classes to simulation.py and remove redundant mesh and thickness properties.
